refactor(tests): 合并重复的测试文件并整理测试代码
本次提交将多个分散的测试文件内容合并到主测试文件中，并删除冗余的测试文件：
1. 将catalog_service_s2s_test.go的测试内容合并到catalog_service_test.go
2. 将resource_service_s2s_test.go的测试内容合并到resource_service_test.go
3. 将logic_view_sql_test.go的测试内容合并到新建的logic_view_test.go
4. 将resource_data_service_additional_test.go的测试内容合并到resource_data_service_test.go
5. 将build_task_normalize_test.go的测试内容合并到build_task_service_test.go
6. 删除raw_query_service_internal_test.go并将测试内容迁移到raw_query_service_test.go

同时统一了测试代码的格式和断言风格，使用 testify 库替换原生测试断言，提升代码可读性和维护性。
